Fix constexpr error in utils_web.h
String literals are not guaranteed to be merged by the compiler, so
the result of "foo" == "foo" is undefined and thus not a constexpr.
Older gcc did not catch this, but gcc7 now rightfully errors out. As a
band-aid, change it to use strcmp(), which is not required by the
standard to be a constexpr for string constants, but it is a constexpr
with gcc. The proper fix would be to use an enum for error codes instead
of strings.