Clean config REST endpoint
This is a refactoring of the /api/v1/admin/config REST endpoint,
allowing the API to store arbitrary JSON objects and removing most
hard-coded logic from the endpoint.

Due to unfortunate timing issues, there are a couple of points to
address as soon as possible:
* lobby upstream czmq for marking zconfig_remove() as stable
* update 42ity's czmq with upstream's czmq
* remove local zconfig_remove() implementation hack
* split mapping data into separate files inside file-system
* prune dead code

This modification should be backwards-compatible.

Signed-off-by: Jean-Baptiste Boric <Jean-BaptisteBORIC@Eaton.com>