Revert "auth_verify.ecpp : pretend user.reauth(true) if session is OK and this lowest-priority auth implem ends up being THE handler"
This reverts commit c6d5c18371795a9f929cbc464e42e20428997985.

Seems the evaluation that this is a no-op if higher priority auth verifiers
are present was wrong, and this servlet does get called after all. So this
hardcoding of "reauth happened" should happen either conditionally, based
on some system setting or detected presence of those other verifiers,
or earlier in stack / per default setting of the flag, and flipped by
others if deemed fit. Maybe a tri-state value, for definitive knowledge
about flag state, is better than a bool here.